+++ resolwe/test/utils.py
@@ -74,7 +74,6 @@ def with_docker_executor(method):
         executor_settings = settings.FLOW_EXECUTOR.copy()
         executor_settings.update({
             'NAME': 'resolwe.flow.executors.docker',
-            'CONTAINER_IMAGE': 'resolwe/test:base'
         })
 
         try: